deletion of alias document structures.

+++ document_structure.go
@@ -323,10 +323,18 @@ func deleteDocumentStructure(w http.ResponseWriter, r *http.Request) {
     return
   }
 
-  _, err = SQLDB.Exec("delete from qf_fields where dsid = ?", dsid)
+  isAlias, _, err := DSIdAliasPointsTo(ds)
   if err != nil {
     errorPage(w, err.Error())
     return
+  }
+
+  if ! isAlias {
+    _, err = SQLDB.Exec("delete from qf_fields where dsid = ?", dsid)
+    if err != nil {
+      errorPage(w, err.Error())
+      return
+    }
   }
 
   _, err = SQLDB.Exec("delete from qf_permissions where dsid = ?", dsid)
